+++ package.json
@@ -18,8 +18,8 @@
     "date-fns": "^3.4.0",
     "gray-matter": "^4.0.3",
     "i": "^0.3.7",
-    "next": "^15.2.4",
     "micro": "^10.0.1",
+    "next": "^15.2.4",
     "next-auth": "^4.24.11",
     "next-themes": "^0.3.0",
     "nextjs-toploader": "^3.7.15",
@@ -38,12 +38,12 @@
   },
   "devDependencies": {
     "@types/aos": "^3.0.7",
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@types/node": "20.17.30",
+    "@types/react": "18.3.20",
     "@types/react-dom": "^18",
     "@types/react-slick": "^0.23.13",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "5.8.2"
   }
 }
